move logging to utils
