--- docs/api/logging.md
@@ -1,14 +1,0 @@
-# Logging
-
-Helper functions for logging during use of package.
-
-```{eval-rst}
-.. currentmodule:: crested
-```
-
-```{eval-rst}
-.. autosummary::
-    :toctree: _autosummary
-
-    setup_logging
-```
