+++ docs/api/utils.md
@@ -14,4 +14,5 @@ CREsted provides a few utility function to help with sequence encoding, function
     extract_bigwig_values_per_bp
     hot_encoding_to_sequence
     one_hot_encode_sequence
+    setup_logging
 ```
