+++ src/crested/__init__.py
@@ -23,7 +23,6 @@ from importlib.metadata import version
 from . import pl, pp, tl, utils
 from ._datasets import get_dataset, get_motif_db
 from ._io import import_beds, import_bigwigs
-from ._logging import setup_logging
 
 __all__ = [
     "pl",
@@ -32,7 +31,6 @@ __all__ = [
     "utils",
     "import_beds",
     "import_bigwigs",
-    "setup_logging",
     "get_dataset",
     "get_motif_db",
 ]
@@ -42,4 +40,4 @@ __version__ = version("crested")
 os.environ["AUTOGRAPH_VERBOSITY"] = "0"
 
 # Setup loguru logging
-setup_logging(log_level="INFO", log_file=None)
+utils.setup_logging(log_level="INFO", log_file=None)
