+++ src/crested/pl/hist/_distribution.py
@@ -8,8 +8,8 @@ import seaborn as sns
 from anndata import AnnData
 from loguru import logger
 
-from crested._logging import log_and_raise
 from crested.pl._utils import render_plot
+from crested.utils._logging import log_and_raise
 
 
 def distribution(
